Added package watcher to allow gin to listen to changes in packages
The package watcher looks at the package in the current working directory,
and parses the source code to find out the imports. It then figures out
which of those imports are outside of the GOROOT (and thus not part of the stdlib),
and uses fsnotify.v1 to listen to changes to files in all those packages.

This allows gin to; for example, reload the web server providing a web API to another
library that is not a subpackage of the web server.