Config.Laddr default value localhost
When creating a listener with an empty address, the listener will assume any interface (0.0.0.0). This will trigger firewall permissions to be granted every time a unique binary is generated. This can be problematic during tests as the binary has a unique name by default.

Since this is a proxy for development sake. It is safe to assume if Config.Laddr is not provided we can use localhost to avoid the firewall module.